+++ KeireClient/Include/KeireClient/Editor/AssetBrowserPanel.h
@@ -40,6 +40,7 @@ namespace KeireEditor
         virtual bool CreateAssetBrowserAnimationGraph(std::string_view name) = 0;
         virtual bool CreateAssetBrowserProceduralMotionProfile(std::string_view name) = 0;
         virtual bool CreateAssetBrowserScript(std::string_view name) = 0;
+        virtual bool CreateAssetBrowserScriptableObjectScript(std::string_view name) = 0;
         virtual bool CreateAssetBrowserManagedAssembly(std::string_view name) = 0;
         virtual bool CreateAssetBrowserManagedData(Keire::ManagedTypeId type, std::string_view name) = 0;
         virtual bool CreateAssetBrowserAudioMixer(std::string_view name) = 0;