+++ KeireClient/Include/KeireClient/Editor/ShaderGraphPanel.h
@@ -161,6 +161,7 @@ namespace KeireEditor
         bool m_PreviewDirty = false;
         bool m_OwnJobSystem = false;
         bool m_NodeMenuOpen = false;
+        bool m_OpenFunctionExtractionPopup = false;
         bool m_ShowPreview = true;
     };
 } // namespace KeireEditor